+++ youkai-console-java/src/main/java/com/github/kazuhitom/youkaicheckdigittry/YoukaiTest03.java
@@ -1,8 +1,9 @@
 package com.github.kazuhitom.youkaicheckdigittry;
 
-import com.github.kazuhitom.youkaicheckdigittry.state.A31F;
-import com.github.kazuhitom.youkaicheckdigittry.state.AttackCharacters;
-import com.github.kazuhitom.youkaicheckdigittry.converter.CodeToCharacterConverter;
+import com.github.kazuhitom.youkaicheckdigittry.checkdigit.YoukaiPasswordAttacker;
+import com.github.kazuhitom.youkaicheckdigittry.checkdigit.state.A31F;
+import com.github.kazuhitom.youkaicheckdigittry.checkdigit.state.AttackCharacters;
+import com.github.kazuhitom.youkaicheckdigittry.checkdigit.converter.CodeToCharacterConverter;
 import sun.misc.Signal;
 
 import java.io.PrintStream;
