+++ youkai-console-java/src/main/java/com/github/kazuhitom/youkaicheckdigittry/checkdigit/converter/CodeToCharacterConverter.java
@@ -1,4 +1,4 @@
-package com.github.kazuhitom.youkaicheckdigittry.converter;
+package com.github.kazuhitom.youkaicheckdigittry.checkdigit.converter;
 
 import java.util.stream.IntStream;
