+++ youkai-console-java/src/test/java/com/github/kazuhitom/youkaicheckdigittry/state/AttackCharactersTest.java
@@ -1,6 +1,7 @@
 package com.github.kazuhitom.youkaicheckdigittry.state;
 
-import com.github.kazuhitom.youkaicheckdigittry.converter.CodeToCharacterConverter;
+import com.github.kazuhitom.youkaicheckdigittry.checkdigit.converter.CodeToCharacterConverter;
+import com.github.kazuhitom.youkaicheckdigittry.checkdigit.state.AttackCharacters;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
